pyproject.toml
[project]
name = "lotl_detector"
version = "0.1.0"
description = "Efficient LOTL (Living Off The Land) Attack Detector"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "sentence-transformers>=2.2.0",
    "chainlit>=1.0.0",
    "shap>=0.42.0",
    "torch>=1.12.0; platform_system != 'darwin' or platform_machine != 'arm64'",
    "torch>=2.0.0; platform_system == 'darwin' and platform_machine == 'arm64'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]

